samd/Pin: Change the Pin naming scheme.
Pin numbers are now the MCU port numbers in the range

PA0..PA31:  0..31
PB0..PB31: 32..63
PC0..PC31: 64..95
PD0..PD31: 96..127

Pins can be denoted by the GPIO port number, the
name as defined in pins.csv or a string in the
form: Pxnn, like "PA16" or "PD03".

The pins.c and pins.h files are now obsolete. The
pin objects are part of the AF table.

As result of a simplification, the code now supports
uisng pin names or numbers instead of pin objects
for modules like UART, SPI, PWM, I2C, ADC, pininfo.