rp2: Prevent use of BT and Wifi.
Due to an open stability issue with the chip firmware, heavy use of BT
and Wifi at the same time causes a corruption of chip memory. Until this
is fixed, prevent the use of BT and Wifi at the same time. Add the hooks
necessary to implement this.

For AP mode, make connect/disconnect work like
active(True)/active(False)

For STA mode make connect call active(True) if not done already.