py/mpthread: Add RTOS-support function to run on mp thread.
This abstracts the implementation currently used by modbluetooth.c
to get BLE events from the host stack running in an RTOS task into
a MicroPython task.

This is currently only implemented when the GIL is enabled, but it should
also be possible to make this work using the scheduler as a fallback.

This work was funded through GitHub Sponsors.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>