unix/mpbthciport: Use mp_thread_run_on_mp_thread.
This replaces the previous code that used the scheduler (via thread) to run
the BLE host stack. Now the host stack can run directly on the thread, and
then modbluetooth.c will use mp_thread_run_on_mp_thread to run the
Python-level callbacks on a MicroPython thread as necessary.

Also implements the HCI layer more like an RTOS port, where you have a task
running the host stack, and UART data is sent directly to NimBLE from a
UART IRQ (which in this case is another thread poll()'ing directly on the
UART). This makes it behave more like ESP32 or Zephyr, where we only run
the Python-level callbacks in a MicroPython thread.

This work was funded through GitHub Sponsors.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>