extmod/nimble/nimble/nimble_npl_os: Implement synchronisation.
This provides a full implementation of the NimBLE mutex, critical section,
and semaphore. On ports without threading, this is a no-op change, but this
is necessary to make NimBLE's locking functionality work on threaded ports.

NimBLE also requires a recursive mutex. Rather than implementing low-level
support for a recursive mutex (e.g. in each port's mpthreadport.c) we
provide a basic implementation at this layer that works for NimBLE's
requirements.

This work was funded through GitHub Sponsors.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>