tests/multi_bluetooth: Improve event ordering handling.
For _IRQ_GATTC_READ_DONE, we should wait for it to ensure that we receive
it before we sent the write (this prevents an ordering race with the
"gattc_write" vs the "_IRQ_GATTC_READ_DONE" line).

For _IRQ_GATTC_WRITE_DONE, don't print it out because we wait for it anyway
(so it will be confirmed to have been generated), and it can race with
the concurrent notify or indicate event.

This work was funded through GitHub Sponsors.

Signed-off-by: Jim Mussared <jim.mussared@gmail.com>